+++ lib/exmbus/parser/apl/full_frame.ex
@@ -3,56 +3,62 @@ defmodule Exmbus.Parser.Apl.FullFrame do
   Parser for a full frame.
 
   A FullFrame contains a list of data records.
+  The parser expects the remaining binary to be the APL records,
+  and will recursively call `DataRecord.parse/1` until it runs out of bytes.
   """
 
-  alias Exmbus.Parser.Context
-  alias Exmbus.Parser.Apl.DataRecord.InvalidDataRecord
   alias Exmbus.Parser.Apl.DataRecord
+  alias Exmbus.Parser.Apl.DataRecord.InvalidDataRecord
   alias Exmbus.Parser.Apl.FormatFrame
+  alias Exmbus.Parser.Context
 
   defstruct records: [],
             manufacturer_bytes: nil
 
   def parse(%Context{} = ctx) do
-    parse_full_frame(ctx.bin, ctx, [])
+    # recursively parse records into the accumulator.
+    # we consume the entire binary from the context here,
+    # since we expect to be able to handle all of it internally.
+    parse_records(%{ctx | bin: <<>>}, ctx.bin, [])
   end
 
-  defp parse_full_frame(<<>>, ctx, acc) do
-    finalize_full_frame(<<>>, ctx, acc)
+  # when we run out of bytes, we finalize the FullFrame parse
+  defp parse_records(ctx, <<>>, acc) do
+    finalize(ctx, <<>>, acc)
   end
 
-  defp parse_full_frame(bin, ctx, acc) do
+  defp parse_records(%{} = ctx, bin, acc) do
     case DataRecord.parse(bin, ctx) do
       {:ok, %DataRecord{} = data_record, rest} ->
-        parse_full_frame(rest, ctx, [data_record | acc])
+        parse_records(ctx, rest, [data_record | acc])
 
       {:ok, %InvalidDataRecord{} = data_record, rest} ->
-        parse_full_frame(rest, ctx, [data_record | acc])
+        parse_records(ctx, rest, [data_record | acc])
 
       # just skip the idle filler
       {:special_function, :idle_filler, rest} ->
-        parse_full_frame(rest, ctx, acc)
+        parse_records(ctx, rest, acc)
 
       # manufacturer specific data is the rest of the APL data
       {:special_function, {:manufacturer_specific, :to_end}, rest} ->
-        finalize_full_frame(rest, ctx, acc)
+        finalize(ctx, rest, acc)
 
       {:special_function, {:manufacturer_specific, :more_records_follow}, rest} ->
-        finalize_full_frame(rest, ctx, acc)
+        finalize(ctx, rest, acc)
 
       {:error, _reason, _rest} = e ->
         e
     end
   end
 
-  defp finalize_full_frame(rest, %{} = ctx, acc) do
+  defp finalize(%{} = ctx, rest, acc) do
     full_frame = %__MODULE__{
       records: Enum.reverse(acc),
       # all remaining bytes are manufacturer specific:
       manufacturer_bytes: rest
     }
 
-    {:next, %{ctx | bin: <<>>, apl: full_frame}}
+    {:next, %{ctx | apl: full_frame}}
   end
 
   def format_signature(%__MODULE__{} = ff) do